requirements.txt
Django==1.8
django-appconf==1.0.2
django-avatar==4.0.1
django-bootstrap-form==3.2.1
django-bootstrap-forms==0.1
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.6.1
django-djconfig==0.5.3
django-filter==1.1.0
django-forms-bootstrap==3.1.0
django-haystack==2.7.dev0
django-infinite-scroll-pagination==0.2.0
django-material==1.0.0
django-picklefield==1.0.0
django-suit==0.2.25
django-uni-form==0.9.0
django-viewflow==1.0.2
mistune==0.7.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.1
python-magic==0.4.13
pytz==2017.2
six==1.11.0
South==1.0.2
uni-slugify==0.1.4
Unidecode==0.4.21
Whoosh==2.7.4
XlsxWriter==0.9.7
